use Gate to enhance the authorisation to Tag action

+++ app/Http/Controllers/HobbyTagController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use App\Tag;
 use App\Hobby;
 
@@ -24,6 +25,11 @@ class HobbyTagController extends Controller
     public function attachTag($hobby_id, $tag_id)
     {
         $hobby = Hobby::find($hobby_id);
+
+        if(Gate::denies('connect_hobby_tag', $hobby)){
+            abort(403, 'Unauthorised User');
+        }
+
         $tag = Tag::find($tag_id);
         $hobby->tags()->attach($tag_id);
         return back()->with(
@@ -36,6 +42,11 @@ class HobbyTagController extends Controller
     public function detachTag($hobby_id, $tag_id)
     {
         $hobby = Hobby::find($hobby_id);
+
+        if(Gate::denies('connect_hobby_tag', $hobby)){
+            abort(403, 'Unauthorised User');
+        }
+
         $tag = Tag::find($tag_id);
         $hobby->tags()->detach($tag_id);
         return back()->with(
